+++ app/data/test-data.js
@@ -1370,6 +1370,9 @@ var formWithUserData =
     {"questionCode": "q4", "dataType": "DT", "header": false, "units": null, "codingInstructions": null, "value": "Wed Nov 17 2015 00:00:00 GMT-0500 (EST)",
       "questionCardinality": null, "answerCardinality": null, "question": "With data type DT", "answers": null,
       "skipLogic": null, "editable": null, "defaultAnswer": null, "displayControl": null, "calculationMethod": null, "items": null},
+    {"questionCode": "q99", "dataType": "DTM", "header": false, "units": null, "codingInstructions": null, "value": "Wed Nov 20 2015 10:10:10 GMT-0500 (EST)",
+      "questionCardinality": null, "answerCardinality": null, "question": "With data type DTM", "answers": null,
+      "skipLogic": null, "editable": null, "defaultAnswer": null, "displayControl": null, "calculationMethod": null, "items": null},
     // answer list, default answer
     {"questionCode": "q5", "dataType": "CNE", "header": false, "units": null, "codingInstructions": null,
       "questionCardinality": null, "answerCardinality": null, "question": "With data type CNE, default value", "answers": [