+++ app/scripts/fhir/R4/sdc-export.js
@@ -14,7 +14,7 @@
  * convertLFormsToQuestionnaireResponse()
  * -- Generate FHIR (standard or SDC) QuestionnaireResponse data from captured data in LForms
  */
-var sdcVersion = '3.5.0';
+var sdcVersion = '2.7';
 var fhirVersionNum = '4.0';
 
 var self = {
@@ -324,24 +324,10 @@ var self = {
           // }
           case "minExclusive":
           case "minInclusive":
-            if (dataType === "DT" || dataType === "DTM" || dataType === "TM" ||
-                dataType === "REAL" || dataType === "INT" ) {
-              extValue = {
-                "url":"http://hl7.org/fhir/StructureDefinition/minValue"
-              };
-              extValue[valueKey] = parseInt(value);
-            }
-            break;
           // http://hl7.org/fhir/StructureDefinition/maxValue
           case "maxExclusive":
           case "maxInclusive":
-            if (dataType === "DT" || dataType === "DTM" || dataType === "TM" ||
-                dataType === "REAL" || dataType === "INT" ) {
-              extValue = {
-                "url":"http://hl7.org/fhir/StructureDefinition/maxValue"
-              };
-              extValue[valueKey] = parseInt(value);
-            }
+            extValue = this._exportMinMax(dataType, value, valueKey, key);
             break;
           // http://hl7.org/fhir/StructureDefinition/minLength
           case "minLength":