+++ app/scripts/fhir/STU3/sdc-import.js
@@ -725,6 +725,9 @@ function addSDCImportFns(ns) {
             }
             break;
           case "DT":
+            item.value = qrValue.valueDate;
+            break;
+          case "DTM":
             item.value = qrValue.valueDateTime;
             break;
           case "CNE":