+++ test/protractor/spec/R4/fhir_context.js
@@ -86,20 +86,21 @@ module.exports = function mockFHIRContext(fhirVersion, weightQuantity) {
                 "text": "Not at all"
               }
               break;
+            default:
+              entry = null;
+              break;
           }
 
           return {
             then: function(callback) {
+              var data = {
+                "resourceType": "Bundle",
+                "type": "searchset",
+              };
+              if (entry)
+                data.entry = [entry];
               setTimeout(function() { // preserve expected async behavior
-                callback({
-                  "data": {
-                    "resourceType": "Bundle",
-                    "type": "searchset",
-                    "entry": [
-                      entry
-                    ]
-                  }
-                });
+                callback({data: data});
               });
             }
           };