+++ CHANGELOG.md
@@ -2,10 +2,37 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [17.2.0] 2019-06-27
+### Added
+- Added support for multiple code system in answers.
+
+## [17.1.1] 2019-06-26
+### Fixed
+- Fixed a bug in FHIR/SDC Questionnaire version detection.
+
+## [17.1.0] 2019-06-20
+### Added
+- DTM support and datetime picker
+
+## [17.0.0] 2019-06-21
+### Changed (for FHIR forms only)
+- This contains a breaking change that only affects forms imported from FHIR
+  Questionnaires.  The loading of external FHIR resources no longer happens when
+  an LFormsData object is constructed.  Instead, after constructing LFormsData,
+  one must call (on the instance) loadFHIRResources(), which will return a
+  promise that resolves when the resources are loaded.  Similarly, if using
+  LForms.Util.addFormToPage(), that will also return a promise that resolves
+  when external FHIR resources had been loaded.  (See the
+  [documentation](https://lhncbc.github.io/lforms/index.html#fhirQueries).)
+
+## [16.0.1] 2019-06-18
+### Fixed
+- Fixed extension URLs for SDC Questionnaire
+
 ## [16.0.0] 2019-06-17
 ### Fixed
-- Each value of the multiple selected values in CNE/CWE typed item 
-  should be a separate Observation resource when LForms data is converted 
+- Each value of the multiple selected values in CNE/CWE typed item
+  should be a separate Observation resource when LForms data is converted
   to corresponding FHIR resources.
 
 ## [15.7.2] 2019-06-14